refactor(test-run): split pipeline responsibilities for cohesion
Why:
- test run 基盤で Service / Results / Artifacts / Process 実行の責務が混在し、凝集をさらに高める必要があった。

What Changed:
- ProcessRunner 契約と実装を TestRun 配下から Execution 共通層へ移動した。
- TestRunService を Preflight / Pipeline / ResultMapper の分割構造に再設計した。
- UnityResultsConverter を XML parser と artifact writer に分割した。
- TestRunArtifactsService から meta.json 書き込みとファイル契約定数を分離した。
- 関連テストの参照先を更新し、既存挙動が維持されることを確認した。

Impact:
- test run の主要コンポーネントが単一責務に近づき、変更理由ごとの修正範囲が限定される。
- 将来の Process 実行再利用や results/artifacts 契約変更を局所化できる。
