+++ src/Ucli/TestRun/Artifacts/TestRunArtifactFileNames.cs
@@ -1,0 +1,20 @@
+namespace MackySoft.Ucli.TestRun.Artifacts;
+
+/// <summary> Defines fixed file names used in one test-run artifacts directory. </summary>
+internal static class TestRunArtifactFileNames
+{
+    /// <summary> Gets the metadata file name. </summary>
+    public const string MetaJson = "meta.json";
+
+    /// <summary> Gets the Unity raw test-results XML file name. </summary>
+    public const string ResultsXml = "results.xml";
+
+    /// <summary> Gets the Unity editor log file name. </summary>
+    public const string EditorLog = "editor.log";
+
+    /// <summary> Gets the normalized test-results JSON file name. </summary>
+    public const string ResultsJson = "results.json";
+
+    /// <summary> Gets the summary JSON file name. </summary>
+    public const string SummaryJson = "summary.json";
+}
