+++ src/Ucli/TestRun/Artifacts/TestRunMetaStore.cs
@@ -1,0 +1,83 @@
+using System.Globalization;
+using System.Text.Json;
+using MackySoft.Ucli.TestRun.Configuration;
+
+namespace MackySoft.Ucli.TestRun.Artifacts;
+
+/// <summary> Implements metadata JSON writing for test-run artifact sessions. </summary>
+internal sealed class TestRunMetaStore : ITestRunMetaStore
+{
+    private const int MetaSchemaVersion = 1;
+
+    private static readonly JsonSerializerOptions SerializerOptions = new()
+    {
+        PropertyNamingPolicy = JsonNamingPolicy.CamelCase,
+        WriteIndented = false,
+    };
+
+    /// <summary> Writes one metadata snapshot for a run session. </summary>
+    /// <param name="configuration"> The resolved test-run configuration. </param>
+    /// <param name="session"> The artifacts session. </param>
+    /// <param name="finishedAtUtc"> The completion timestamp to persist. </param>
+    public void Write (
+        ResolvedTestRunConfiguration configuration,
+        ArtifactsSession session,
+        DateTimeOffset finishedAtUtc)
+    {
+        ArgumentNullException.ThrowIfNull(configuration);
+        ArgumentNullException.ThrowIfNull(session);
+
+        var payload = new MetaJsonPayload(
+            SchemaVersion: MetaSchemaVersion,
+            RunId: session.RunId,
+            StartedAt: session.StartedAtUtc.ToString("O", CultureInfo.InvariantCulture),
+            FinishedAt: finishedAtUtc.ToString("O", CultureInfo.InvariantCulture),
+            ProjectPath: configuration.UnityProject.UnityProjectRoot,
+            UnityVersion: configuration.UnityVersion,
+            UnityEditorPath: configuration.UnityEditorPath,
+            Mode: configuration.Mode,
+            TestPlatform: TestRunPlatformCodec.ToValue(configuration.TestPlatform),
+            BuildTarget: configuration.BuildTarget,
+            TestFilter: configuration.TestFilter,
+            TestCategories: configuration.TestCategories,
+            AssemblyNames: configuration.AssemblyNames,
+            TestSettingsPath: configuration.TestSettingsPath,
+            ArtifactsDir: session.ArtifactsDir);
+
+        var json = JsonSerializer.Serialize(payload, SerializerOptions);
+        File.WriteAllText(session.Paths.MetaJsonPath, json);
+    }
+
+    /// <summary> Represents metadata payload for one test-run artifacts session. </summary>
+    /// <param name="SchemaVersion"> The metadata schema version. </param>
+    /// <param name="RunId"> The run identifier. </param>
+    /// <param name="StartedAt"> The run start timestamp in ISO-8601 UTC format. </param>
+    /// <param name="FinishedAt"> The run completion timestamp in ISO-8601 UTC format. </param>
+    /// <param name="ProjectPath"> The Unity project path. </param>
+    /// <param name="UnityVersion"> The resolved Unity version. </param>
+    /// <param name="UnityEditorPath"> The resolved Unity editor path. </param>
+    /// <param name="Mode"> The execution mode option value. </param>
+    /// <param name="TestPlatform"> The test-platform value. </param>
+    /// <param name="BuildTarget"> The optional build target value. </param>
+    /// <param name="TestFilter"> The optional test-filter value. </param>
+    /// <param name="TestCategories"> The normalized test-category values. </param>
+    /// <param name="AssemblyNames"> The normalized assembly-name values. </param>
+    /// <param name="TestSettingsPath"> The optional test-settings path value. </param>
+    /// <param name="ArtifactsDir"> The run artifacts directory path. </param>
+    private sealed record MetaJsonPayload (
+        int SchemaVersion,
+        string RunId,
+        string StartedAt,
+        string FinishedAt,
+        string ProjectPath,
+        string UnityVersion,
+        string UnityEditorPath,
+        string Mode,
+        string TestPlatform,
+        string? BuildTarget,
+        string? TestFilter,
+        string[] TestCategories,
+        string[] AssemblyNames,
+        string? TestSettingsPath,
+        string ArtifactsDir);
+}
